pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aac"
version = "3.6.0"
description = "Accelerated Arbitrage Corp — AI-driven multi-exchange trading system"
readme = "README.md"
requires-python = ">=3.9,<3.14"  # aiohttp hangs on 3.14+
license = { text = "Proprietary" }
authors = [
    { name = "Resonance Energy Corp", email = "acceleratedarbitrage@outlook.com" },
    { name = "Resonance Energy Corp", email = "gripandripphdd@outlook.com" },
]
keywords = ["trading", "arbitrage", "fintech", "crypto", "ai", "quantitative"]

# ── Core runtime dependencies ──────────────────────────────────────────────────
dependencies = [
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "ccxt>=4.2.0",
    "yfinance>=0.2.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "web3>=6.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "xgboost>=2.0.0",
    "structlog>=23.0.0",
    "rich>=13.0.0",
    "dash>=2.14.0",
    "plotly>=5.17.0",
    "dash-bootstrap-components>=1.5.0",
    "praw>=7.7.0",
    "apscheduler>=3.10.0",
    "textblob>=0.17.0",
    "kafka-python>=2.0.2",
    "redis>=5.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "pydantic>=2.4.0",
    "cryptography>=42.0.0",
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.0",
    "pyotp>=2.6.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
# ML extras (large downloads — optional)
ml = [
    "torch>=2.1.0",
    "transformers>=4.35.0",
    "gymnasium>=0.29.0",
]

# Visualization
viz = [
    "matplotlib>=3.7.0",
    "seaborn>=0.13.0",
]

# Dev tools
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-timeout>=2.2.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "ruff>=0.1.0",
]

# Production extras
prod = [
    "prometheus-client>=0.18.0",
    "confluent-kafka>=2.3.0",
    "hiredis>=2.3.0",
]

[project.scripts]
# Entry points — run from command line after `pip install -e .`
aac            = "launch:main"   # Unified launcher (replaces all below)
aac-launch     = "core.aac_master_launcher:main"
aac-dashboard  = "monitoring.aac_master_monitoring_dashboard:main"
aac-paper      = "trading.paper_trading_validation:main"
aac-setup      = "scripts.setup_production:main"

[tool.setuptools.packages.find]
where = ["."]
include = [
    "shared*",
    "strategies*",
    "TradingExecution*",
    "BigBrainIntelligence*",
    "CentralAccounting*",
    "CryptoIntelligence*",
    "monitoring*",
    "trading*",
    "integrations*",
    "core*",
    "deployment*",
    "agents*",
    "tools*",
    "demos*",
    "reddit*",
    "models*",
    "docs*",
]
exclude = [
    "tests*",
    "archive*",
    "version_control*",
    "aac*",
    "build*",
    "dist*",
]

# ── Tool configs ───────────────────────────────────────────────────────────────

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
extend-exclude = '''
/(
    \.venv
  | archive
  | version_control
  | aac/aac069
  | aac/NCC
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_glob = [".venv/*", "archive/*", "version_control/*", "aac/aac069/*"]

[tool.ruff]
line-length = 100
target-version = "py39"
exclude = [".venv", "archive", "version_control", "aac/aac069"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = [
    "E402",   # module-level import not at top (sys.path inserts)
    "E501",   # line-length handled by formatter
    "E711",   # comparison to None
    "E712",   # comparison to True/False
    "E721",   # type comparison (use isinstance)
    "E722",   # bare except
    "E731",   # lambda assignment
    "F401",   # unused imports (many used at runtime / re-exports)
    "F403",   # wildcard imports
    "F405",   # may be undefined from wildcard import
    "F541",   # f-string without placeholders
    "F811",   # redefinition of unused name
    "F841",   # unused local variable
    "UP",     # pyupgrade (all rules — not critical)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]  # re-exports
"tests/*" = ["F401", "F811"]       # fixtures

[tool.pylint.messages_control]
disable = [
    "C0114",  # missing-module-docstring
    "C0115",  # missing-class-docstring
    "C0116",  # missing-function-docstring
    "C0209",  # consider-using-f-string
    "C0301",  # line-too-long
    "C0303",  # trailing-whitespace
    "C0304",  # missing-final-newline
    "C0411",  # wrong-import-order
    "R0801",  # duplicate-code
    "R0903",  # too-few-public-methods
    "R0913",  # too-many-arguments
    "R0914",  # too-many-locals
    "W0107",  # unnecessary-pass
    "W0201",  # attribute-defined-outside-init
    "W0611",  # unused-import
    "W0612",  # unused-variable
    "W0613",  # unused-argument
    "W0621",  # redefined-outer-name
    "W0622",  # redefined-builtin
    "W0718",  # broad-exception-caught
    "W1203",  # logging-fstring-interpolation
    "E1101",  # no-member (too many false positives)
]

[tool.pylint.format]
max-line-length = 100

[tool.pyright]
typeCheckingMode = "off"
venvPath = "."
venv = ".venv"
reportMissingTypeStubs = false
reportRedefinition = false

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
exclude = [".venv/", "archive/", "version_control/", "aac/aac069/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "live: requires live API credentials",
    "integration: requires external service",
    "slow: runtime > 5s",
    "paper: paper-trading mode",
    "exchange: requires exchange API",
]

[tool.bandit]
exclude_dirs = [".venv", "archive", "tests", "docs", "build"]
skips = ["B101", "B311"]  # B101=assert, B311=random (not crypto-grade)
